Update Dashboard component to conditionally render NFTProjectsPanel

+++ src/app/services/dashboardService.ts
@@ -1,8 +1,16 @@
 import { NFTProject } from "@/types/NFTProject";
+import { useWalletStore } from "../store/walletStore";
 
 // call GET projects from the API
 export const getProjects = async () => {
-  const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}projects`);
+  const jwt = useWalletStore.getState().jwt;
+  if (!jwt) return [];
+  const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}projects`, {
+    headers: {
+      Authorization: `Bearer ${jwt}`,
+    },
+  });
+
   if (!response.ok) {
     throw new Error("Error in the response");
   }
@@ -15,10 +23,15 @@ export const getProjects = async () => {
 };
 
 export const deleteProject = async (id: string) => {
+  const jwt = useWalletStore.getState().jwt;
+  if (!jwt) return [];
   const response = await fetch(
     `${process.env.NEXT_PUBLIC_API_URL}projects/${id}`,
     {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${jwt}`,
+      },
     }
   );
   if (!response.ok) {
@@ -27,10 +40,13 @@ export const deleteProject = async (id: string) => {
 };
 
 export const addProject = async (project: NFTProject) => {
+  const jwt = useWalletStore.getState().jwt;
+  if (!jwt) return [];
   const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}projects`, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
+      Authorization: `Bearer ${jwt}`,
     },
     body: JSON.stringify(project),
   });
